refactor: remove metric mesos (#1411)
* bugfix: fix issue #1301

* refactor: remove metric_mesos logic

* refactor: update metric urls

* minor: remove Metric model logic
