--- bcs-app/backend/container_service/observability/metric/open_apis/__init__.py
@@ -1,1 +1,0 @@
-# -*- coding: utf-8 -*-
