+++ bcs-app/backend/container_service/observability/metric/views/__init__.py
@@ -16,6 +16,7 @@ from .cluster import ClusterMetricViewSet  # noqa
 from .container import ContainerMetricViewSet  # noqa
 from .node import NodeMetricViewSet  # noqa
 from .pod import PodMetricViewSet  # noqa
+from .prometheus import PrometheusUpdateViewSet  # noqa
 from .service import ServiceViewSet  # noqa
 from .service_monitor import ServiceMonitorDetailViewSet, ServiceMonitorViewSet  # noqa
 from .target import TargetsViewSet  # noqa
