--- bcs-app/backend/container_service/observability/metric_mesos/admin.py
@@ -1,25 +1,0 @@
-# -*- coding: utf-8 -*-
-"""
-Tencent is pleased to support the open source community by making 蓝鲸智云PaaS平台社区版 (BlueKing PaaS Community
-Edition) available.
-Copyright (C) 2017-2021 THL A29 Limited, a Tencent company. All rights reserved.
-Licensed under the MIT License (the "License"); you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-
-    http://opensource.org/licenses/MIT
-
-Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on
-an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the
-specific language governing permissions and limitations under the License.
-"""
-from django.contrib import admin
-
-from backend.container_service.observability.metric_mesos.models import Metric
-
-
-class MetricAdmin(admin.ModelAdmin):
-    list_display = ('project_id', 'name', 'port', 'uri', 'frequency', 'version', 'data_id', 'metric_type', 'timeout')
-    search_fields = ('project_id', 'name', 'port', 'uri', 'frequency', 'version', 'data_id', 'metric_type', 'timeout')
-
-
-admin.site.register(Metric, MetricAdmin)
