--- bcs-app/backend/container_service/observability/metric_mesos/urls.py
@@ -1,29 +1,0 @@
-# -*- coding: utf-8 -*-
-"""
-Tencent is pleased to support the open source community by making 蓝鲸智云PaaS平台社区版 (BlueKing PaaS Community
-Edition) available.
-Copyright (C) 2017-2021 THL A29 Limited, a Tencent company. All rights reserved.
-Licensed under the MIT License (the "License"); you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-
-    http://opensource.org/licenses/MIT
-
-Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on
-an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the
-specific language governing permissions and limitations under the License.
-"""
-from django.conf.urls import url
-
-from . import views
-
-urlpatterns = [
-    url(r'^api/metric/(?P<project_id>\w{32})/$', views.collector.Metric.as_view({'get': 'list', 'post': 'create'})),
-    url(
-        r'^api/metric/(?P<project_id>\w{32})/(?P<metric_id>\d+)/$',
-        views.collector.MetricDetail.as_view({'get': 'get', 'put': 'put', 'post': 'recreate', 'delete': 'delete'}),
-    ),
-    url(
-        r'^api/metric/(?P<project_id>\w{32})/instances/(?P<metric_id>\d+)/$',
-        views.collector.MetricIns.as_view({'get': 'get_instance'}),
-    ),
-]
